Repair test failure caused by test double
A classic example of how test doubles can hurt you. This test was
failing because I assumed that TypeMeta would be set on the fake
Service object managed by test infrastructure (as it would be in
a real k8s environment). But this isn't so, so when the test went
looking for that information, it failed.